requirements.txt
django
google-genai
requests
python-dotenv
django-cors-headers
djangorestframework